add built js files for 6.0.0
