+++ src/components/organisms/header-nav/components/burger-menu.js
@@ -9,15 +9,13 @@ Object.defineProperty(exports, "__esModule", {
 });
 exports.BurgerMenu = void 0;
 
-var _objectWithoutProperties2 = _interopRequireDefault(require("@babel/runtime/helpers/objectWithoutProperties"));
-
 var _extends2 = _interopRequireDefault(require("@babel/runtime/helpers/extends"));
+
+var _objectWithoutProperties2 = _interopRequireDefault(require("@babel/runtime/helpers/objectWithoutProperties"));
 
 var _react = _interopRequireWildcard(require("react"));
 
 var _burgerIcon = require("../../../../components/graphics/icons/burger-icon");
-
-var _dropdown = require("./dropdown");
 
 var _visuallyHidden = require("../../../../components/accessibility/visually-hidden");
 
@@ -27,56 +25,46 @@ var _context = require("./context");
 
 var _classnames = _interopRequireDefault(require("classnames"));
 
+var _useDropdown2 = require("../hooks/use-dropdown");
+
+var _dropdownButton = require("./dropdown-button");
+
 var namespace = 'kkbbAndCo';
-var DROPDOWN_CLASS = "".concat(namespace, "-PlatformMenu");
 var CLOSE_EVENT = "".concat(namespace, ":platformMenu:close");
-
-var ButtonIcon = function ButtonIcon(props) {
-  return /*#__PURE__*/_react.default.createElement(_react.default.Fragment, null, /*#__PURE__*/_react.default.createElement(_burgerIcon.BurgerIcon, (0, _extends2.default)({}, props, {
-    hoverColor: _colorsConfig.default.font1,
-    className: "k-ButtonIcon__svg",
-    "aria-hidden": "true"
-  })), /*#__PURE__*/_react.default.createElement(_visuallyHidden.VisuallyHidden, null, "Menu"));
-};
-
-var buttonClassNames = (0, _classnames.default)('k-HeaderNav__BurgerMenu__button', 'k-ButtonIcon', 'k-ButtonIcon--tiny');
 
 var BurgerMenu = function BurgerMenu(_ref) {
   var children = _ref.children,
       dropdownContentWidth = _ref.dropdownContentWidth,
-      props = (0, _objectWithoutProperties2.default)(_ref, ["children", "dropdownContentWidth"]);
-  var dropdownComponent = (0, _react.useRef)(null);
+      className = _ref.className,
+      props = (0, _objectWithoutProperties2.default)(_ref, ["children", "dropdownContentWidth", "className"]);
 
-  var getElementById = function getElementById(id) {
-    return function () {
-      return document.getElementById(id);
-    };
-  };
+  var _useContext = (0, _react.useContext)(_context.Context),
+      id = _useContext.id,
+      callOnToggle = _useContext.callOnToggle;
 
-  return /*#__PURE__*/_react.default.createElement(_context.Context.Consumer, null, function (_ref2) {
-    var id = _ref2.id,
-        callOnToggle = _ref2.callOnToggle;
-    return /*#__PURE__*/_react.default.createElement(_dropdown.Dropdown, (0, _extends2.default)({}, props, {
-      buttonClassName: buttonClassNames,
-      buttonContentOnCollapsed: /*#__PURE__*/_react.default.createElement(ButtonIcon, {
-        isAnimatedOnHover: true
-      }),
-      buttonContentOnExpanded: /*#__PURE__*/_react.default.createElement(ButtonIcon, {
-        isActive: true
-      }),
-      buttonId: "".concat(id, "PlateformMenu"),
-      className: DROPDOWN_CLASS,
-      closeEvents: [CLOSE_EVENT],
-      closeOnOuterClick: true,
-      dropdownContent: children,
-      dropdownContentWidth: dropdownContentWidth,
-      onToggle: callOnToggle,
-      positionedVerticallyWith: getElementById(id),
-      positionedWithBorder: true,
-      ref: dropdownComponent,
-      refreshEvents: ['resize']
-    }));
-  });
+  var _useDropdown = (0, _useDropdown2.useDropdown)({
+    dropdownContentWidth: dropdownContentWidth,
+    callOnToggle: callOnToggle,
+    dropdownAnchorSide: 'left',
+    closeEvents: [CLOSE_EVENT],
+    buttonId: "".concat(id, "PlateformMenu"),
+    menuId: "".concat(id, "PlateformMenu__content")
+  }),
+      dropdownProps = _useDropdown.dropdownProps,
+      buttonProps = _useDropdown.buttonProps,
+      menuProps = _useDropdown.menuProps,
+      isDropdownExpanded = _useDropdown.isDropdownExpanded;
+
+  return /*#__PURE__*/_react.default.createElement("div", (0, _extends2.default)({}, dropdownProps, props, {
+    className: (0, _classnames.default)(className, dropdownProps.className)
+  }), /*#__PURE__*/_react.default.createElement(_dropdownButton.DropdownButton, (0, _extends2.default)({}, buttonProps, {
+    className: (0, _classnames.default)(buttonProps.className, 'k-HeaderNav__BurgerMenu__button', 'k-ButtonIcon', 'k-ButtonIcon--tiny')
+  }), /*#__PURE__*/_react.default.createElement(_burgerIcon.BurgerIcon, {
+    isActive: isDropdownExpanded,
+    hoverColor: _colorsConfig.default.font1,
+    className: "k-ButtonIcon__svg",
+    "aria-hidden": "true"
+  }), /*#__PURE__*/_react.default.createElement(_visuallyHidden.VisuallyHidden, null, "Menu")), /*#__PURE__*/_react.default.createElement("div", menuProps, children));
 };
 
 exports.BurgerMenu = BurgerMenu;
