--- src/components/organisms/header-nav/components/user-menu-next.js
@@ -1,127 +1,0 @@
-"use strict";
-
-var _interopRequireWildcard = require("@babel/runtime/helpers/interopRequireWildcard");
-
-var _interopRequireDefault = require("@babel/runtime/helpers/interopRequireDefault");
-
-Object.defineProperty(exports, "__esModule", {
-  value: true
-});
-exports.UserMenuNext = void 0;
-
-var _extends2 = _interopRequireDefault(require("@babel/runtime/helpers/extends"));
-
-var _toConsumableArray2 = _interopRequireDefault(require("@babel/runtime/helpers/toConsumableArray"));
-
-var _objectWithoutProperties2 = _interopRequireDefault(require("@babel/runtime/helpers/objectWithoutProperties"));
-
-var _react = _interopRequireWildcard(require("react"));
-
-var _propTypes = _interopRequireDefault(require("prop-types"));
-
-var _context = require("./context");
-
-var _reactElements = require("../../../../helpers/react/react-elements");
-
-var _classnames = _interopRequireDefault(require("classnames"));
-
-var _dropdownButton = require("./dropdown-button");
-
-var _useDropdown2 = require("../hooks/use-dropdown");
-
-var _arrowIcon = require("../../../../components/graphics/icons/arrow-icon");
-
-var namespace = 'kkbbAndCo';
-var CLOSE_EVENT = "".concat(namespace, ":userMenu:close");
-
-var UserMenuNext = function UserMenuNext(_ref) {
-  var children = _ref.children,
-      dropdownContentWidth = _ref.dropdownContentWidth,
-      closeEvents = _ref.closeEvents,
-      dropdownAnchorSide = _ref.dropdownAnchorSide,
-      className = _ref.className,
-      padding = _ref.padding,
-      mobilePadding = _ref.mobilePadding,
-      props = (0, _objectWithoutProperties2.default)(_ref, ["children", "dropdownContentWidth", "closeEvents", "dropdownAnchorSide", "className", "padding", "mobilePadding"]);
-
-  var _useContext = (0, _react.useContext)(_context.Context),
-      id = _useContext.id,
-      callOnToggle = _useContext.callOnToggle;
-
-  var buttonChild = (0, _reactElements.getReactElementsByType)({
-    children: children,
-    type: UserMenuNext.Button
-  })[0];
-  var menuChild = (0, _reactElements.getReactElementsByType)({
-    children: children,
-    type: UserMenuNext.Navigation
-  })[0];
-
-  var _useDropdown = (0, _useDropdown2.useDropdown)({
-    dropdownContentWidth: dropdownContentWidth,
-    callOnToggle: callOnToggle,
-    dropdownAnchorSide: dropdownAnchorSide,
-    closeEvents: [CLOSE_EVENT].concat((0, _toConsumableArray2.default)(closeEvents)),
-    buttonId: "".concat(id, "__UserMenu__Button"),
-    menuId: "".concat(id, "__UserMenu__Menu")
-  }),
-      dropdownProps = _useDropdown.dropdownProps,
-      buttonProps = _useDropdown.buttonProps,
-      menuProps = _useDropdown.menuProps,
-      isDropdownExpanded = _useDropdown.isDropdownExpanded;
-
-  return /*#__PURE__*/_react.default.createElement("div", (0, _extends2.default)({}, dropdownProps, props, {
-    className: (0, _classnames.default)(className, dropdownProps.className)
-  }), /*#__PURE__*/_react.default.createElement(_dropdownButton.DropdownButton, (0, _extends2.default)({}, buttonProps, {
-    style: {
-      '--UserMenu-Button-backgroundColor': buttonChild.props.backgroundColor || null,
-      '--UserMenu-Button-backgroundColorHover': buttonChild.props.backgroundColorHover || null,
-      '--UserMenu-Button-backgroundColorActive': buttonChild.props.backgroundColorActive || null,
-      '--UserMenu-Button-color': buttonChild.props.color || null,
-      '--UserMenu-Button-colorHover': buttonChild.props.colorHover || null,
-      '--UserMenu-Button-colorActive': buttonChild.props.colorActive || null
-    },
-    className: (0, _classnames.default)('k-HeaderNav__UserMenuButton', buttonChild.props.className, buttonProps.className, {
-      'k-HeaderNav__UserMenuButton--hasArrow': buttonChild.props.hasArrow,
-      'k-HeaderNav__UserMenuButton--noPadding': !padding,
-      'k-HeaderNav__UserMenuButton--noPaddingMobile': !mobilePadding
-    })
-  }), (0, _react.cloneElement)(buttonChild), buttonChild.props.hasArrow && /*#__PURE__*/_react.default.createElement(_arrowIcon.ArrowIcon, {
-    direction: isDropdownExpanded ? 'top' : 'bottom',
-    className: "k-u-margin-left-single k-u-hidden@xs-down",
-    color: "currentColor"
-  })), /*#__PURE__*/_react.default.createElement("div", menuProps, (0, _react.cloneElement)(menuChild)));
-};
-
-exports.UserMenuNext = UserMenuNext;
-
-UserMenuNext.Button = function (_ref2) {
-  var children = _ref2.children;
-  return /*#__PURE__*/_react.default.createElement(_react.default.Fragment, null, children);
-};
-
-UserMenuNext.Navigation = function (_ref3) {
-  var children = _ref3.children;
-  return /*#__PURE__*/_react.default.createElement(_react.default.Fragment, null, children);
-};
-
-UserMenuNext.propTypes = {
-  dropdownContentWidth: _propTypes.default.oneOfType([_propTypes.default.number, _propTypes.default.string]),
-  padding: _propTypes.default.bool,
-  mobilePadding: _propTypes.default.bool,
-  closeEvents: _propTypes.default.arrayOf(_propTypes.default.string),
-  dropdownAnchorSide: _propTypes.default.oneOf(['left', 'right'])
-};
-UserMenuNext.defaultProps = {
-  dropdownContentWidth: null,
-  padding: true,
-  mobilePadding: true,
-  closeEvents: [],
-  dropdownAnchorSide: 'left'
-};
-UserMenuNext.Button.propTypes = {
-  hasArrow: _propTypes.default.bool
-};
-UserMenuNext.Button.defaultProps = {
-  hasArrow: false
-};
