+++ src/components/organisms/header-nav/components/user-menu.js
@@ -19,13 +19,17 @@ var _react = _interopRequireWildcard(require("react"));
 
 var _propTypes = _interopRequireDefault(require("prop-types"));
 
-var _dropdown = require("./dropdown");
-
 var _context = require("./context");
 
 var _reactElements = require("../../../../helpers/react/react-elements");
 
 var _classnames = _interopRequireDefault(require("classnames"));
+
+var _dropdownButton = require("./dropdown-button");
+
+var _useDropdown2 = require("../hooks/use-dropdown");
+
+var _arrowIcon = require("../../../../components/graphics/icons/arrow-icon");
 
 var namespace = 'kkbbAndCo';
 var CLOSE_EVENT = "".concat(namespace, ":userMenu:close");
@@ -33,97 +37,79 @@ var CLOSE_EVENT = "".concat(namespace, ":userMenu:close");
 var UserMenu = function UserMenu(_ref) {
   var children = _ref.children,
       dropdownContentWidth = _ref.dropdownContentWidth,
-      padding = _ref.padding,
       closeEvents = _ref.closeEvents,
-      buttonProps = _ref.buttonProps,
+      dropdownAnchorSide = _ref.dropdownAnchorSide,
       className = _ref.className,
-      props = (0, _objectWithoutProperties2.default)(_ref, ["children", "dropdownContentWidth", "padding", "closeEvents", "buttonProps", "className"]);
-  var userDropdownRef = (0, _react.useRef)(null);
+      padding = _ref.padding,
+      mobilePadding = _ref.mobilePadding,
+      props = (0, _objectWithoutProperties2.default)(_ref, ["children", "dropdownContentWidth", "closeEvents", "dropdownAnchorSide", "className", "padding", "mobilePadding"]);
 
-  var getElementById = function getElementById(id) {
-    return function () {
-      return document.getElementById(id);
-    };
-  };
+  var _useContext = (0, _react.useContext)(_context.Context),
+      id = _useContext.id,
+      callOnToggle = _useContext.callOnToggle;
 
-  var getButtonId = function getButtonId(id) {
-    return "".concat(id, "UserMenu");
-  };
-
-  var button = (0, _reactElements.getReactElementsByType)({
+  var buttonChild = (0, _reactElements.getReactElementsByType)({
     children: children,
     type: UserMenu.Button
   })[0];
-  var navigation = (0, _reactElements.getReactElementsByType)({
+  var menuChild = (0, _reactElements.getReactElementsByType)({
     children: children,
     type: UserMenu.Navigation
   })[0];
-  var _button$props = button.props,
-      hasArrow = _button$props.hasArrow,
-      backgroundColor = _button$props.backgroundColor,
-      backgroundColorHover = _button$props.backgroundColorHover,
-      backgroundColorActive = _button$props.backgroundColorActive,
-      color = _button$props.color,
-      colorHover = _button$props.colorHover,
-      colorActive = _button$props.colorActive;
-  var dropdownClassName = (0, _classnames.default)('k-HeaderNav__UserMenu', "".concat(namespace, "-UserMenu"), className);
-  var buttonStyles = {
-    '--UserMenu-Button-backgroundColor': backgroundColor,
-    '--UserMenu-Button-backgroundColorHover': backgroundColorHover,
-    '--UserMenu-Button-backgroundColorActive': backgroundColorActive,
-    '--UserMenu-Button-color': color,
-    '--UserMenu-Button-colorHover': colorHover,
-    '--UserMenu-Button-colorActive': colorActive
-  };
-  var buttonClassName = (0, _classnames.default)('k-HeaderNav__UserMenuButton', {
-    'k-HeaderNav__UserMenuButton--hasArrow': hasArrow,
-    'k-HeaderNav__UserMenuButton--noPadding': !padding
-  });
-  return /*#__PURE__*/_react.default.createElement(_context.Context.Consumer, null, function (_ref2) {
-    var id = _ref2.id,
-        callOnToggle = _ref2.callOnToggle;
-    return /*#__PURE__*/_react.default.createElement(_dropdown.Dropdown, (0, _extends2.default)({}, props, {
-      button: button,
-      buttonClassName: buttonClassName,
-      buttonStyles: buttonStyles,
-      buttonId: getButtonId(id),
-      className: dropdownClassName,
-      closeEvents: [CLOSE_EVENT].concat((0, _toConsumableArray2.default)(closeEvents)),
-      closeOnOuterClick: true,
-      dropdownContent: navigation,
-      dropdownContentWidth: dropdownContentWidth,
-      hasArrow: hasArrow,
-      keepInitialButtonAction: true,
-      onToggle: callOnToggle,
-      positionedHorizontallyWith: getElementById(getButtonId(id)),
-      positionedVerticallyWith: getElementById(id),
-      ref: userDropdownRef,
-      refreshEvents: ['resize']
-    }));
-  });
+
+  var _useDropdown = (0, _useDropdown2.useDropdown)({
+    dropdownContentWidth: dropdownContentWidth,
+    callOnToggle: callOnToggle,
+    dropdownAnchorSide: dropdownAnchorSide,
+    closeEvents: [CLOSE_EVENT].concat((0, _toConsumableArray2.default)(closeEvents)),
+    buttonId: "".concat(id, "__UserMenu__Button"),
+    menuId: "".concat(id, "__UserMenu__Menu")
+  }),
+      dropdownProps = _useDropdown.dropdownProps,
+      buttonProps = _useDropdown.buttonProps,
+      menuProps = _useDropdown.menuProps,
+      isDropdownExpanded = _useDropdown.isDropdownExpanded;
+
+  return /*#__PURE__*/_react.default.createElement("div", (0, _extends2.default)({}, dropdownProps, props, {
+    className: (0, _classnames.default)('k-HeaderNav__UserMenu', className, dropdownProps.className)
+  }), /*#__PURE__*/_react.default.createElement(_dropdownButton.DropdownButton, (0, _extends2.default)({}, buttonProps, {
+    style: {
+      '--UserMenu-Button-backgroundColor': buttonChild.props.backgroundColor || null,
+      '--UserMenu-Button-backgroundColorHover': buttonChild.props.backgroundColorHover || null,
+      '--UserMenu-Button-backgroundColorActive': buttonChild.props.backgroundColorActive || null,
+      '--UserMenu-Button-color': buttonChild.props.color || null,
+      '--UserMenu-Button-colorHover': buttonChild.props.colorHover || null,
+      '--UserMenu-Button-colorActive': buttonChild.props.colorActive || null
+    },
+    className: (0, _classnames.default)('k-HeaderNav__UserMenuButton', buttonChild.props.className, buttonProps.className)
+  }), (0, _react.cloneElement)(buttonChild), /*#__PURE__*/_react.default.createElement(_arrowIcon.ArrowIcon, {
+    direction: isDropdownExpanded ? 'top' : 'bottom',
+    className: "k-u-margin-left-single k-u-hidden@xs-down",
+    color: "currentColor"
+  })), /*#__PURE__*/_react.default.createElement("div", menuProps, (0, _react.cloneElement)(menuChild)));
 };
 
 exports.UserMenu = UserMenu;
-UserMenu.propTypes = {
-  dropdownContentWidth: _propTypes.default.oneOfType([_propTypes.default.number, _propTypes.default.string]),
-  padding: _propTypes.default.bool,
-  closeEvents: _propTypes.default.arrayOf(_propTypes.default.string),
-  hasArrow: _propTypes.default.bool
-};
-UserMenu.defaultProps = {
-  dropdownContentWidth: null,
-  padding: true,
-  closeEvents: [],
-  hasArrow: false
+
+UserMenu.Button = function (_ref2) {
+  var children = _ref2.children;
+  return /*#__PURE__*/_react.default.createElement(_react.default.Fragment, null, children);
 };
 
-UserMenu.Button = function (_ref3) {
+UserMenu.Navigation = function (_ref3) {
   var children = _ref3.children;
   return /*#__PURE__*/_react.default.createElement(_react.default.Fragment, null, children);
 };
 
-UserMenu.Navigation = function (_ref4) {
-  var children = _ref4.children,
-      props = (0, _objectWithoutProperties2.default)(_ref4, ["children"]);
-  return /*#__PURE__*/_react.default.createElement("div", props, children);
+UserMenu.propTypes = {
+  dropdownContentWidth: _propTypes.default.oneOfType([_propTypes.default.number, _propTypes.default.string]),
+  padding: _propTypes.default.bool,
+  mobilePadding: _propTypes.default.bool,
+  closeEvents: _propTypes.default.arrayOf(_propTypes.default.string),
+  dropdownAnchorSide: _propTypes.default.oneOf(['left', 'right'])
+};
+UserMenu.defaultProps = {
+  dropdownContentWidth: null,
+  closeEvents: [],
+  dropdownAnchorSide: 'left'
 };
