+++ src/esm/components/organisms/header-nav/config.js
@@ -1,3 +1,1 @@
-export var MOBILE_HEADER_HEIGHT = 65;
-export var TABLET_HEADER_HEIGHT = 80;
-export var DESKTOP_HEADER_HEIGHT = 80;
+export var HEADER_HEIGHT = 65;
