+++ src/esm/components/organisms/header-nav/styles.js
@@ -1,11 +1,10 @@
 import styled from 'styled-components';
-import { ScreenConfig } from '../../../constants/screen-config';
+import { ScreenConfig, mq } from '../../../constants/screen-config';
 import TYPOGRAPHY from '../../../constants/typography-config';
 import COLORS from '../../../constants/colors-config';
-import { CONTAINER_PADDING } from '../../../constants/grid-config';
 import { pxToRem, stepToRem } from '../../../helpers/utils/typography';
-import { MOBILE_HEADER_HEIGHT, TABLET_HEADER_HEIGHT, DESKTOP_HEADER_HEIGHT } from './config';
+import { HEADER_HEIGHT } from './config';
 export var StyledHeader = styled.header.withConfig({
   displayName: "styles__StyledHeader",
   componentId: "sc-1h9gzly-0"
-})(["position:relative;z-index:var(--HeaderNav-zIndex);.k-HeaderNav__stickyContainer{width:100%;}&.k-HeaderNav--menuIsExpanded{z-index:var(--HeaderNav-zIndex-openMenu);.k-HeaderNav__stickyContainer{transition:none;}}&.k-HeaderNav--shadow{.k-Spacer + .k-HeaderNav__stickyContainer{box-shadow:0 ", " ", " rgba(0,0,0,0.1);}}&.k-HeaderNav--border .k-HeaderNav__stickyContainer{box-shadow:0 ", " 0 rgba(0,0,0,0.08);}.k-HeaderNav{display:flex;align-items:center;width:100%;box-sizing:border-box;background:", ";transition:background-color 0.2s ease;@media (min-width:", "px){padding:0 ", " 0 ", ";}&:not(.k-HeaderNav--kkbb__only){@media (min-width:", "px){padding:0;}}.quickAccessLink{background:", ";transition:background-color 0.2s ease;height:100%;}&,*{", ";box-sizing:border-box;}*:not(input):focus{outline-offset:", ";outline-width:", ";outline-style:solid;outline-color:", ";}*:not(input):focus:not(:focus-visible){outline-color:transparent;}*:not(input):focus-visible{outline-color:", ";}}&.k-HeaderNav--inactiveBackground .k-HeaderNav,&.k-HeaderNav--inactiveBackground .k-HeaderNav__Button,&.k-HeaderNav--inactiveBackground .quickAccessLink{background-color:", ";}.k-HeaderNav__BurgerMenu__button{padding:0;height:100%;@media (min-width:", "px){padding:0 ", " 0 ", ";}}.k-HeaderNav__QuickAccessLink{position:absolute;display:inline-flex;align-items:center;left:-100%;top:0;bottom:0;z-index:var(--HeaderNav_quickAccessLink_zIndex,2);padding:0 ", ";background-color:", ";border-right:", " solid ", ";", " color:", ";line-height:1;font-size:", ";text-decoration:none;transition:all 0.2s ease;transition-delay:0,0;opacity:0;&:focus,&:active{left:0;opacity:1;transition-delay:0,0.2s;}}.k-HeaderNav__Logo{padding:0;display:flex;align-items:center;@media (min-width:", "px){padding:", ";}}.k-HeaderNav__Button{", " box-sizing:border-box;display:flex;align-items:center;justify-content:center;gap:", ";height:100%;font-size:", ";color:var(--HeaderMenu-Button-color);background-color:var(--HeaderMenu-Button-backgroundColor);text-decoration:none;transition:background-color 0.2s ease,color 0.2s ease;&:focus{outline:", " solid ", ";outline-offset:", ";}&:focus:not(:focus-visible){outline-color:transparent;}&:focus-visible{outline-color:", ";}&:hover,&:focus{color:var( --HeaderMenu-Button-colorHover,var(--HeaderMenu-Button-color) );background-color:var(--HeaderMenu-Button-backgroundColorHover);}.k-HeaderNav__Button__text:first-child{margin-right:", ";}}.k-HeaderNav__right,.k-HeaderNav__centered{display:flex;height:100%;align-items:center;}.k-HeaderNav__right{justify-content:flex-end;flex-grow:1;&.k-HeaderNav__right--padded{padding-right:", ";gap:", ";.k-HeaderNav__Button:last-child{margin-right:", ";}@media (min-width:", "px){padding-right:", ";gap:", ";.k-HeaderNav__Button:last-child{margin-right:", ";}}}}.k-HeaderNav__centered{justify-content:center;flex-grow:100;}.k-HeaderNav__column{flex-direction:column;}.k-HeaderNav-nav{list-style-type:none;display:flex;padding:0;margin:0;height:100%;}.k-HeaderNav__UserMenu{height:100%;display:flex;.k-Avatar__text.k-Avatar__text--hasEllipsis{max-width:", ";}}.k-HeaderNav__UserMenuButton{align-self:center;display:flex;align-items:center;justify-content:center;height:100%;border:0;font-size:", ";line-height:normal;background-color:", ";background-color:var( --UserMenu-Button-backgroundColor,", " );.k-Badge{border-color:", ";border-color:var( --UserMenu-Button-backgroundColor,", " );}&,.k-Avatar__text{color:", ";color:var(--UserMenu-Button-color,", ");transition:background-color 0.2s ease,color 0.2s ease;}.k-Avatar{align-self:center;}@media (max-width:", "){padding:0 ", ";}&:hover{cursor:pointer;background-color:", ";background-color:var( --UserMenu-Button-backgroundColorHover,", " );.k-Badge{border-color:", ";border-color:var( --UserMenu-Button-backgroundColorHover,", " );}&,.k-Avatar__text{color:", ";color:var(--UserMenu-Button-colorHover,", ");}}}.k-HeaderNav__UserMenuButton[aria-expanded='true'],.k-Dropdown--isExpanded .k-HeaderNav__UserMenuButton{&,&:hover{background-color:", ";background-color:var( --UserMenu-Button-backgroundColorActive,", " );.k-Badge{border-color:", ";border-color:var( --UserMenu-Button-backgroundColorActive,", " );}&,.k-Avatar__text{color:", ";color:var(--UserMenu-Button-colorActive,", ");}}}.k-HeaderNav__UserMenuButton--hasArrow{padding-right:", ";}.k-HeaderNav__UserMenuButton--noPadding{padding:0 !important;}@media (max-width:", "){.k-HeaderNav__UserMenuButton--noPaddingMobile{padding:0 !important;}}.k-HeaderNav-nav--center{justify-content:center;}.k-HeaderNav-nav__element{-ms-overflow-x:hidden;display:flex;align-items:stretch;}.k-HeaderNav-nav__element--fixedSize{flex-shrink:0;}.k-HeaderNav-nav__item{display:flex;align-items:center;border-top:", " solid transparent;", ";font-size:", ";line-height:normal;color:", ";text-align:center;text-decoration:none;cursor:pointer;transition:color 0.2s,border-color 0.2s;&:hover,&:focus{color:", ";text-decoration:none;}&:active{border-color:currentColor;color:", ";}&.is-selected{border-color:currentColor;color:", ";}}&.k-HeaderNav--small .k-HeaderNav{height:", ";.k-HeaderNav__UserMenuButton{padding:0 ", ";&.k-Dropdown__button,&.k-HeaderNav__Button{min-width:", ";}}.k-HeaderNav-nav__item{padding:0 ", ";}.k-Dropdown__button,.k-HeaderNav__Button{min-width:", ";cursor:pointer;}.k-HeaderNav__Button--hasText{.k-HeaderNav__Button__text{&:first-child{margin-left:", ";}&:last-child{margin-right:", ";}}}}&.k-HeaderNav--regular .k-HeaderNav{height:", ";@media (min-width:", "px){height:", ";}@media (min-width:", "px){height:", ";}.k-HeaderNav__UserMenuButton{padding:0 ", ";}.k-HeaderNav-nav__item{padding:0 ", ";&.smallPadding{padding:0 ", ";}}.k-HeaderNav__right .k-HeaderNav-nav__item{padding:0 ", ";}.k-Dropdown__button,.k-HeaderNav__Button{min-width:", ";cursor:pointer;@media (min-width:", "px){min-width:", ";}@media (min-width:", "px){min-width:", ";}}.k-HeaderNav__Button__text:first-child{margin-left:", ";@media (min-width:", "px){margin-left:", ";}}.k-HeaderNav__Button--smallPadding{.k-HeaderNav__Button__text:first-child{margin-left:0;}}.k-HeaderNav__Button__text:last-child{margin-right:", ";@media (min-width:", "px){margin-right:", ";}}.k-HeaderNav__Button--smallPadding{.k-HeaderNav__Button__text:last-child{margin-left:", ";margin-right:0;}}}.k-Dropdown{display:flex;align-self:stretch;}.k-Dropdown__content{position:absolute;top:0;z-index:20;min-width:", ";margin-top:", ";visibility:hidden;opacity:0;transition:margin 0.2s,visibility 0.2s,opacity 0.2s;max-height:calc(100vh - 100%);box-shadow:0 ", " ", " rgba(0,0,0,0.1);@media (max-width:", "){min-width:0;width:100% !important;left:0 !important;right:0 !important;}}.k-Dropdown--isExpanded .k-Dropdown__content{margin-top:0;visibility:visible;opacity:1;overflow-x:hidden;overflow-y:scroll;}.k-HeaderNavDropdown{align-self:stretch;.k-HeaderNavDropdown__menu{position:absolute;top:100%;z-index:20;max-height:calc(100vh - 100%);box-shadow:0 ", " ", " rgba(0,0,0,0.1);transition:margin 0.2s,visibility 0.2s,opacity 0.2s;margin-top:", ";visibility:hidden;opacity:0;}@media (max-width:", "){.k-HeaderNavDropdown__menu{min-width:100vw;max-width:100vw;left:0;right:0;}}@media (min-width:", "){position:relative;.k-HeaderNavDropdown__menu{min-width:max(", ",100%);&.k-HeaderNavDropdown__menu--is-left{left:0;}&.k-HeaderNavDropdown__menu--is-right{right:0;}}}&.k-HeaderNavDropdown--isExpanded .k-HeaderNavDropdown__menu{margin-top:0;visibility:visible;opacity:1;overflow-x:hidden;overflow-y:scroll;}}@media (min-width:", "){.k-HeaderNav__right--padded .k-HeaderNavDropdown__menu{margin-left:", ";margin-right:", ";}}.k-HeaderNav__searchInput{position:relative;transition:max-width 0.2s ease;height:100%;display:flex;align-items:center;justify-content:flex-end;> *{flex:1 1 auto;}.k-Form-TextInput{text-overflow:ellipsis;overflow:hidden;", ";width:100%;}@media (max-width:", "){flex:0 0 auto;&.k-HeaderNav__searchInput--mobileInvisible .k-Form-TextInputWithButton{display:none;}.k-HeaderNav__searchInput__mobileFold{flex:0 0 auto;padding-left:", ";padding-right:", ";display:inline-flex;align-items:center;align-self:stretch;}&:not(.k-HeaderNav__searchInput--mobileInvisible){background:", ";z-index:2;position:absolute;top:0;left:0;right:0;bottom:0;padding-left:", ";}}@media (min-width:", "){flex:1 0 auto;max-width:min(100%,", ");.k-HeaderNav__searchInput__mobileFold{display:none;}}}.k-HeaderNav__floatingDropdown{position:absolute;left:0;right:0;min-width:fit-content;background-color:", ";z-index:20;margin-top:", ";visibility:hidden;opacity:0;transition:margin 0.2s,visibility 0.2s,opacity 0.2s;@media (max-width:", "){top:100%;max-height:calc(100vh - 100%);max-width:100vw;}@media (min-width:", "){top:calc(100% + ", ");max-height:calc(100vh - (100% + ", "));box-shadow:0 ", " ", " rgba(0,0,0,0.1);border-radius:", ";}&.k-HeaderNav__floatingDropdown--isExpanded{margin-top:0;visibility:visible;opacity:1;overflow-x:hidden;overflow-y:scroll;}}"], pxToRem(2), pxToRem(4), pxToRem(1), COLORS.background1, ScreenConfig.L.min, pxToRem(40), pxToRem(30), ScreenConfig.L.min, COLORS.background1, TYPOGRAPHY.fontStyles.regular, pxToRem(-2), pxToRem(2), COLORS.primary4, COLORS.primary4, COLORS.background3, ScreenConfig.S.min, pxToRem(15), pxToRem(CONTAINER_PADDING), pxToRem(30), COLORS.background1, pxToRem(1), COLORS.line1, TYPOGRAPHY.fontStyles.regular, COLORS.font1, pxToRem(16), ScreenConfig.S.min, pxToRem(10), TYPOGRAPHY.fontStyles.regular, pxToRem(10), stepToRem(-1), COLORS.primary4, pxToRem(2), pxToRem(-4), COLORS.primary4, pxToRem(-10), pxToRem(20), pxToRem(10), pxToRem(-10), ScreenConfig.S.min, pxToRem(10), pxToRem(15), pxToRem(-10), pxToRem(145), stepToRem(-1), COLORS.background3, COLORS.background3, COLORS.background3, COLORS.background3, COLORS.font1, COLORS.font1, ScreenConfig.S.max, pxToRem(15), COLORS.line1, COLORS.line1, COLORS.line1, COLORS.line1, COLORS.font1, COLORS.font1, COLORS.background1, COLORS.background1, COLORS.background1, COLORS.background1, COLORS.font1, COLORS.font1, pxToRem(30), pxToRem(ScreenConfig.XS.max), pxToRem(4), TYPOGRAPHY.fontStyles.regular, stepToRem(-1), COLORS.font1, COLORS.primary1, COLORS.primary1, COLORS.primary1, pxToRem(MOBILE_HEADER_HEIGHT), pxToRem(15), pxToRem(40), pxToRem(15), pxToRem(52), pxToRem(15), pxToRem(10), pxToRem(MOBILE_HEADER_HEIGHT), ScreenConfig.S.min, pxToRem(TABLET_HEADER_HEIGHT), ScreenConfig.L.min, pxToRem(DESKTOP_HEADER_HEIGHT), pxToRem(40), pxToRem(40), pxToRem(10), pxToRem(10), pxToRem(52), ScreenConfig.S.min, pxToRem(TABLET_HEADER_HEIGHT), ScreenConfig.L.min, pxToRem(DESKTOP_HEADER_HEIGHT), pxToRem(30), ScreenConfig.S.min, pxToRem(40), pxToRem(30), ScreenConfig.S.min, pxToRem(40), pxToRem(10), pxToRem(200), pxToRem(-10), pxToRem(3), pxToRem(4), pxToRem(ScreenConfig.XS.max), pxToRem(3), pxToRem(4), pxToRem(-10), pxToRem(ScreenConfig.XS.max), pxToRem(ScreenConfig.S.min), pxToRem(200), pxToRem(ScreenConfig.S.min), pxToRem(-15), pxToRem(-15), TYPOGRAPHY.fontStyles.light, pxToRem(ScreenConfig.S.max), pxToRem(8), pxToRem(13), COLORS.background1, pxToRem(20), pxToRem(ScreenConfig.M.min), pxToRem(500), COLORS.background1, pxToRem(-10), pxToRem(ScreenConfig.S.max), pxToRem(ScreenConfig.M.min), pxToRem(10), pxToRem(20), pxToRem(3), pxToRem(4), pxToRem(8));
+})(["position:relative;z-index:var(--HeaderNav-zIndex);.k-HeaderNav__stickyContainer{width:100%;box-shadow:0 ", " 0 rgba(0,0,0,0.08);}&.k-HeaderNav--menuIsExpanded{z-index:var(--HeaderNav-zIndex-openMenu);.k-HeaderNav__stickyContainer{transition:none;}}.k-HeaderNav{display:flex;align-items:center;width:100%;box-sizing:border-box;background:", ";transition:background-color 0.2s ease;height:", ";.quickAccessLink{background:", ";transition:background-color 0.2s ease;height:100%;}&,*{", ";box-sizing:border-box;}*:not(input):focus{outline-offset:", ";outline-width:", ";outline-style:solid;outline-color:", ";}*:not(input):focus:not(:focus-visible){outline-color:transparent;}*:not(input):focus-visible{outline-color:", ";}.k-Dropdown__button,.k-HeaderNav__Button{min-width:", ";cursor:pointer;}}&.k-HeaderNav--inactiveBackground .k-HeaderNav,&.k-HeaderNav--inactiveBackground .k-HeaderNav__Button,&.k-HeaderNav--inactiveBackground .quickAccessLink{background-color:", ";}.k-HeaderNav__BurgerMenu__button{padding:0;height:100%;@media (min-width:", "px){padding:0 ", " 0 ", ";}}.k-HeaderNav__QuickAccessLink{position:absolute;display:inline-flex;align-items:center;left:-100%;top:0;bottom:0;z-index:var(--HeaderNav_quickAccessLink_zIndex,2);padding:0 ", ";background-color:", ";border-right:", " solid ", ";", " color:", ";line-height:1;font-size:", ";text-decoration:none;transition:all 0.2s ease;transition-delay:0,0;opacity:0;&:focus,&:active{left:0;opacity:1;transition-delay:0,0.2s;}}.k-HeaderNav__Logo{padding:0;display:flex;align-items:center;@media ", "{padding:", ";}}@media ", "{.k-HeaderNavDropdown[class*='k-u-hidden@l-up'] + .k-HeaderNav__Logo{padding-left:", ";}}.k-HeaderNav__Button{", " box-sizing:border-box;display:flex;align-items:center;justify-content:center;gap:", ";height:100%;font-size:", ";color:var(--HeaderMenu-Button-color);background-color:var(--HeaderMenu-Button-backgroundColor);text-decoration:none;transition:background-color 0.2s ease,color 0.2s ease;&:focus{outline:", " solid ", ";outline-offset:", ";}&:focus:not(:focus-visible){outline-color:transparent;}&:focus-visible{outline-color:", ";}&:hover,&:focus{color:var( --HeaderMenu-Button-colorHover,var(--HeaderMenu-Button-color) );background-color:var(--HeaderMenu-Button-backgroundColorHover);}&.k-HeaderNav__Button--hasText{.k-HeaderNav__Button__text{&:first-child{margin-left:", ";margin-right:", ";}&:last-child{margin-right:", ";}}}}.k-HeaderNav__right,.k-HeaderNav__centered{display:flex;height:100%;align-items:center;}.k-HeaderNav__right{justify-content:flex-end;flex-grow:1;gap:", ";margin-right:", ";@media ", "{margin-right:", ";}}.k-HeaderNav__centered{justify-content:center;flex-grow:100;}.k-HeaderNav__column{flex-direction:column;}.k-HeaderNav-nav{list-style-type:none;display:flex;padding:0;margin:0;height:100%;}.k-HeaderNav__UserMenu{height:100%;display:flex;.k-Avatar__text.k-Avatar__text--hasEllipsis{max-width:", ";}}.k-HeaderNav__UserMenu{margin-right:", ";@media ", "{margin-right:", ";}}.k-HeaderNav__UserMenuButton{align-self:center;display:flex;align-items:center;justify-content:center;height:100%;border:0;font-size:", ";line-height:normal;background-color:", ";background-color:var( --UserMenu-Button-backgroundColor,", " );padding-right:", ";&.k-Dropdown__button,&.k-HeaderNav__Button{min-width:", ";}.k-Badge{border-color:", ";border-color:var( --UserMenu-Button-backgroundColor,", " );}&,.k-Avatar__text{color:", ";color:var(--UserMenu-Button-color,", ");transition:background-color 0.2s ease,color 0.2s ease;}.k-Avatar{align-self:center;}@media (max-width:", "){padding:0 ", ";}&:hover{cursor:pointer;background-color:", ";background-color:var( --UserMenu-Button-backgroundColorHover,", " );.k-Badge{border-color:", ";border-color:var( --UserMenu-Button-backgroundColorHover,", " );}&,.k-Avatar__text{color:", ";color:var(--UserMenu-Button-colorHover,", ");}}}.k-HeaderNav__UserMenuButton[aria-expanded='true'],.k-Dropdown--isExpanded .k-HeaderNav__UserMenuButton{&,&:hover{background-color:", ";background-color:var( --UserMenu-Button-backgroundColorActive,", " );.k-Badge{border-color:", ";border-color:var( --UserMenu-Button-backgroundColorActive,", " );}&,.k-Avatar__text{color:", ";color:var(--UserMenu-Button-colorActive,", ");}}}.k-HeaderNav-nav--center{justify-content:center;}.k-HeaderNav-nav__element{-ms-overflow-x:hidden;display:flex;align-items:stretch;}.k-HeaderNav-nav__element--fixedSize{flex-shrink:0;}.k-HeaderNav-nav__item{display:flex;align-items:center;padding:0 ", ";border-top:", " solid transparent;border-bottom:", " solid transparent;", ";font-size:", ";line-height:normal;color:", ";text-align:center;text-decoration:none;cursor:pointer;transition:color 0.2s,border-color 0.2s;&:hover,&:focus{color:", ";text-decoration:none;}&:active{border-color:currentColor;color:", ";}&.is-selected{border-top-color:currentColor;color:", ";}}.k-Dropdown{display:flex;align-self:stretch;}.k-Dropdown__content{position:absolute;top:0;z-index:20;min-width:", ";margin-top:", ";visibility:hidden;opacity:0;transition:margin 0.2s,visibility 0.2s,opacity 0.2s;max-height:calc(100vh - 100%);box-shadow:0 ", " ", " rgba(0,0,0,0.1);@media (max-width:", "){min-width:0;width:100% !important;left:0 !important;right:0 !important;}}.k-Dropdown--isExpanded .k-Dropdown__content{margin-top:0;visibility:visible;opacity:1;overflow-x:hidden;overflow-y:scroll;}.k-HeaderNavDropdown{align-self:stretch;.k-HeaderNavDropdown__menu{position:absolute;top:100%;z-index:20;max-height:calc(100vh - 100%);box-shadow:0 ", " ", " rgba(0,0,0,0.1);transition:margin 0.2s,visibility 0.2s,opacity 0.2s;margin-top:", ";visibility:hidden;opacity:0;}@media (max-width:", "){.k-HeaderNavDropdown__menu{min-width:100vw;max-width:100vw;left:0;right:0;}}@media (min-width:", "){position:relative;.k-HeaderNavDropdown__menu{min-width:max(", ",100%);&.k-HeaderNavDropdown__menu--is-left{left:0;}&.k-HeaderNavDropdown__menu--is-right{right:0;}}}&.k-HeaderNavDropdown--isExpanded .k-HeaderNavDropdown__menu{margin-top:0;visibility:visible;opacity:1;overflow-x:hidden;overflow-y:scroll;}}@media (min-width:", "){.k-HeaderNav__right--padded .k-HeaderNavDropdown__menu{margin-left:", ";margin-right:", ";}}.k-HeaderNav__searchInput{position:relative;transition:max-width 0.2s ease;height:100%;display:flex;align-items:center;justify-content:flex-end;> *{flex:1 1 auto;}.k-Form-TextInput{text-overflow:ellipsis;overflow:hidden;", ";width:100%;}@media (max-width:", "){flex:0 0 auto;&.k-HeaderNav__searchInput--mobileInvisible .k-Form-TextInputWithButton{display:none;}.k-HeaderNav__searchInput__mobileFold{flex:0 0 auto;padding-left:", ";padding-right:", ";display:inline-flex;align-items:center;align-self:stretch;}&:not(.k-HeaderNav__searchInput--mobileInvisible){background:", ";z-index:2;position:absolute;top:0;left:0;right:0;bottom:0;padding-left:", ";}}@media (min-width:", "){flex:1 0 auto;max-width:min(100%,", ");.k-HeaderNav__searchInput__mobileFold{display:none;}}}.k-HeaderNav__floatingDropdown{position:absolute;left:0;right:0;min-width:fit-content;background-color:", ";z-index:20;margin-top:", ";visibility:hidden;opacity:0;transition:margin 0.2s,visibility 0.2s,opacity 0.2s;@media (max-width:", "){top:100%;max-height:calc(100vh - 100%);max-width:100vw;}@media (min-width:", "){top:calc(100% + ", ");max-height:calc(100vh - (100% + ", "));box-shadow:0 ", " ", " rgba(0,0,0,0.1);border-radius:", ";}&.k-HeaderNav__floatingDropdown--isExpanded{margin-top:0;visibility:visible;opacity:1;overflow-x:hidden;overflow-y:scroll;}}"], pxToRem(1), COLORS.background1, pxToRem(HEADER_HEIGHT), COLORS.background1, TYPOGRAPHY.fontStyles.regular, pxToRem(-2), pxToRem(2), COLORS.primary4, COLORS.primary4, pxToRem(52), COLORS.background3, ScreenConfig.S.min, pxToRem(10), pxToRem(20), pxToRem(30), COLORS.background1, pxToRem(1), COLORS.line1, TYPOGRAPHY.fontStyles.regular, COLORS.font1, pxToRem(16), mq.tabletAndDesktop, pxToRem(10), mq.desktop, pxToRem(40), TYPOGRAPHY.fontStyles.regular, pxToRem(10), stepToRem(-1), COLORS.primary4, pxToRem(2), pxToRem(-4), COLORS.primary4, pxToRem(15), pxToRem(-10), pxToRem(10), pxToRem(10), pxToRem(20), mq.desktop, pxToRem(40), pxToRem(145), pxToRem(-50), mq.tabletAndDesktop, pxToRem(-40), stepToRem(-1), COLORS.background3, COLORS.background3, pxToRem(50), pxToRem(40), COLORS.background3, COLORS.background3, COLORS.font1, COLORS.font1, ScreenConfig.S.max, pxToRem(15), COLORS.line1, COLORS.line1, COLORS.line1, COLORS.line1, COLORS.font1, COLORS.font1, COLORS.background1, COLORS.background1, COLORS.background1, COLORS.background1, COLORS.font1, COLORS.font1, pxToRem(10), pxToRem(4), pxToRem(4), TYPOGRAPHY.fontStyles.regular, stepToRem(-1), COLORS.font1, COLORS.primary1, COLORS.primary1, COLORS.primary1, pxToRem(200), pxToRem(-10), pxToRem(3), pxToRem(4), pxToRem(ScreenConfig.XS.max), pxToRem(3), pxToRem(4), pxToRem(-10), pxToRem(ScreenConfig.XS.max), pxToRem(ScreenConfig.S.min), pxToRem(250), pxToRem(ScreenConfig.S.min), pxToRem(-15), pxToRem(-15), TYPOGRAPHY.fontStyles.light, pxToRem(ScreenConfig.S.max), pxToRem(8), pxToRem(13), COLORS.background1, pxToRem(20), pxToRem(ScreenConfig.M.min), pxToRem(500), COLORS.background1, pxToRem(-10), pxToRem(ScreenConfig.S.max), pxToRem(ScreenConfig.M.min), pxToRem(10), pxToRem(20), pxToRem(3), pxToRem(4), pxToRem(8));
